Use double precision numbers (#7)

+++ examples/gen_1d.rs
@@ -30,17 +30,17 @@ fn main() {
     let mut output_domain = Domain::new(grid_bound, &mut grid_output);
 
     // Fill in with IC values (use normal dist for spike in the middle)
-    let n_f = buffer_size as f32;
-    let sigma_sq: f32 = (n_f / 25.0) * (n_f / 25.0);
+    let n_f = buffer_size as f64;
+    let sigma_sq: f64 = (n_f / 25.0) * (n_f / 25.0);
     input_domain.par_modify_access(100).for_each(
         |mut d: DomainChunk<'_, GRID_DIMENSION>| {
             d.coord_iter_mut().for_each(
                 |(world_coord, value_mut): (
                     Coord<GRID_DIMENSION>,
-                    &mut f32,
+                    &mut f64,
                 )| {
-                    let x = (world_coord[0] as f32) - (n_f / 2.0);
-                    //let f = ( 1.0 / (2.0 * std::f32::consts::PI * sigma_sq)).sqrt();
+                    let x = (world_coord[0] as f64) - (n_f / 2.0);
+                    //let f = ( 1.0 / (2.0 * std::f64::consts::PI * sigma_sq)).sqrt();
                     let exp = -x * x / (2.0 * sigma_sq);
                     *value_mut = exp.exp()
                 },
